+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Table of Contents
 
+- [v1.41.1](#v1411)
 - [v1.41.0](#v1410)
 - [v1.40.3](#v1403)
 - [v1.40.2](#v1402)
@@ -96,6 +97,13 @@
 - [v0.3.0](#v030)
 - [v0.2.0](#v020)
 - [v0.1.0](#v010)
+
+## [v1.41.1]
+> Release date: 2024/10/22
+
+### Fixes
+- `deck gateway validate` for Konnect supports Konnect configs passed by CLI flags now.
+Earlier, the validation was failing if control plane information was passed via CLI flags.
 
 ## [v1.41.0]
 > Release date: 2024/10/21
@@ -1835,6 +1843,7 @@ No breaking changes have been introduced in this release.
 
 Debut release of decK
 
+[v1.41.1]: https://github.com/Kong/deck/compare/v1.40.0...v1.41.1
 [v1.41.0]: https://github.com/Kong/deck/compare/v1.40.3...v1.41.0
 [v1.40.3]: https://github.com/Kong/deck/compare/v1.40.2...v1.40.3
 [v1.40.2]: https://github.com/Kong/deck/compare/v1.40.1...v1.40.2